Release v3.0.0 and down sizing (#5)
* Release v3.0.0 and down sizing
* Fixed testCreator and package.json
