+++ README-COLLECTION.md
@@ -1,0 +1,15 @@
+# Collection package for `browser-agent`
+
+### Indexed browsers
+
+- Chrome
+- Edge
+- Firefox
+- IE
+- Safari
+
+## How to contribute
+
+Before starting, please read our [Code of Conduct](./CODE_OF_CONDUCT./) and [Contributing Guidelines](./CONTRIBUTING_GUIDELINES.md) for detailed information and instructions on contributing to this repository.
+
+Please note that `browser-agent/collection` is a generated collection from the code-base of `browser-agent`
