+++ README.md
@@ -1,6 +1,6 @@
 # Check Operating System, Browser Type and Version
 
-### Indexed browsers:
+### Browser checks:
 
 - Chrome
 - Edge
@@ -8,7 +8,7 @@
 - IE
 - Safari
 
-### Indexed operating systems:
+### Operating System checks:
 
 - Windows
 - Mac OS
@@ -42,15 +42,21 @@ const shortcutToolTipLabel = (event: Keyboard) => {
 const browserAgent = require('browser-agent')
 const { detectMac } = browserAgent
 
-const shortcutToolTipLabel = (event: Keyboard) => {
+const shortcutToolTipLabel = (event) => {
   const alt = `${detectMac(userAgent) ? '⌥' : 'alt'}`
   const shift = `${detectMac(userAgent) ? '⇧' : 'shift'}`
   return `${alt} + ${shift} + T`
 }
 ```
 
+## Browser Agent packages notice
+
+The npm packages:
+- `browser-agent`
+- `browser-agent/collection`
+
+Have the same version to garuntee compatibility with each other.
+
 ## How to contribute
 
 Before starting, please read our [Code of Conduct](./CODE_OF_CONDUCT./) and [Contributing Guidelines](./CONTRIBUTING_GUIDELINES.md) for detailed information and instructions on contributing to this repository.
-
-
