+++ package.json
@@ -1,9 +1,9 @@
 {
   "name": "browser-agent",
-  "version": "2.0.2",
+  "version": "3.0.0",
   "description": "Get users browser & os information (Uses userAgent)",
   "type": "module",
-  "repository": "git@github.com:TessavWalstijn/browser-id.git",
+  "repository": "git@github.com:TessavWalstijn/browser-agent.git",
   "author": "TessavWalstijn <email@tessavwalstijn.com>",
   "license": "MIT",
   "packageManager": "yarn@4.1.0",
@@ -14,8 +14,11 @@
   },
   "main": "./dist/index.cjs",
   "scripts": {
+    "collection:update": "ts-node --project tsconfig.dev.json ./build/collection",
+    "collection:build": "cd ./collection && tsup ../source/browsers/userAgents.ts --format esm,cjs,iife --dts",
+    "collection:publish": "cd ./collection && npm publish --access public",
     "dev": "ts-node --project tsconfig.dev.json ./source/dev",
-    "build": "tsup source/index.ts --format esm,cjs --dts",
+    "build": "tsup source/index.ts --format esm,cjs,iife --dts",
     "build:test": "vitest --config vitest.build.config.ts",
     "build:tsc": "yarn build:esm &&  yarn build:cjs",
     "build:esm": "npx tsc --build tsconfig.esm.json",
