+++ source/browsers/Chrome/index.ts
@@ -1,2 +1,1 @@
 export * from './chrome.js'
-export * from './chrome.userAgents.js'
