+++ source/browsers/Edge/index.ts
@@ -1,2 +1,1 @@
 export * from './edge.js'
-export * from './edge.userAgents.js'
