+++ source/browsers/Firefox/index.ts
@@ -1,2 +1,1 @@
 export * from './firefox.js'
-export * from './firefox.userAgents.js'
