+++ source/browsers/IE/index.ts
@@ -1,2 +1,1 @@
 export * from './ie.js'
-export * from './ie.userAgents.js'
