+++ source/browsers/Safari/index.ts
@@ -1,2 +1,1 @@
 export * from './safari.js'
-export * from './safari.userAgents.js'
