+++ source/browsers/index.ts
@@ -4,4 +4,3 @@ export * from './Edge/index.js'
 export * from './Firefox/index.js'
 export * from './IE/index.js'
 export * from './Safari/index.js'
-export * from './userAgents.js'
