+++ source/types/userAgent.types.ts
@@ -1,0 +1,10 @@
+import { iBrowser } from './browser.types'
+
+export interface iUserAgent {
+  values: string[]
+  browser: iBrowser
+}
+
+export interface iCollectionUserAgents {
+  [key: string]: iUserAgent[]
+}
